pyproject.toml
[tool.poetry]
name = "drf-api-deploy"
version = "0.1.0"
description = ""
authors = ["farahzuot <farahz997@eng.hu.edu.jo>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.5"
djangorestframework = "^3.12.2"
djangorestframework-simplejwt = "^4.6.0"
psycopg2-binary = "^2.8.6"
django-environ = "^0.4.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
